+++ Apromore-Custom-Plugins/CSVImporter-Logic/pom.xml
@@ -125,11 +125,6 @@
             <artifactId>hadoop-common</artifactId>
             <version>3.3.0</version>
         </dependency>
-        <dependency>
-            <groupId>org.apache.parquet</groupId>
-            <artifactId>parquet-column</artifactId>
-            <version>1.11.0</version>
-        </dependency>
         <!-- https://mvnrepository.com/artifact/org.apache.parquet/parquet-hadoop-bundle -->
         <dependency>
             <groupId>org.apache.parquet</groupId>
@@ -176,7 +171,6 @@
             <artifactId>poi-ooxml</artifactId>
             <version>4.1.2</version>
         </dependency>
-
         <dependency>
             <groupId>com.monitorjbl</groupId>
             <artifactId>xlsx-streamer</artifactId>
